+++ CHANGELOG.md
@@ -7,6 +7,8 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 
+## [1.0.0a5] - 2026-04-08
+
 ### Added
 
 - (Internal) Add new CLI integration tests for more robust testing.
@@ -105,7 +107,8 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - Pre-built plugin: amfi (Mutual Fund)
 - Documentation
 
-[unreleased]: https://github.com/yashovardhan99/niveshpy/compare/v1.0.0a4...HEAD
+[unreleased]: https://github.com/yashovardhan99/niveshpy/compare/v1.0.0a5...HEAD
+[1.0.0a5]: https://github.com/yashovardhan99/niveshpy/compare/v1.0.0a4...v1.0.0a5
 [1.0.0a4]: https://github.com/yashovardhan99/niveshpy/compare/v1.0.0a3...v1.0.0a4
 [1.0.0a3]: https://github.com/yashovardhan99/niveshpy/compare/v1.0.0a2...v1.0.0a3
 [1.0.0a2]: https://github.com/yashovardhan99/niveshpy/compare/v1.0.0a1...v1.0.0a2